+++ src/main/java/org/sopt/api/portfolio/controller/PortfolioController.java
@@ -2,7 +2,7 @@ package org.sopt.api.portfolio.controller;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.sopt.api.portfolio.controller.dto.GetPortfolioDto;
+import org.sopt.api.portfolio.controller.dto.response.GetPortfolioResponseDto;
 import org.sopt.api.portfolio.service.PortfolioService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -15,7 +15,7 @@ public class PortfolioController {
 
     private final PortfolioService portfolioService;
     @RequestMapping
-    public ResponseEntity<List<GetPortfolioDto>> getPortfolio(){
+    public ResponseEntity<List<GetPortfolioResponseDto>> getPortfolio(){
         return ResponseEntity.ok(portfolioService.getPortfolio());
     }
 }
