+++ src/main/java/org/sopt/api/portfolio/controller/dto/response/GetPortfolioResponseDto.java
@@ -1,19 +1,18 @@
-package org.sopt.api.portfolio.controller.dto;
+package org.sopt.api.portfolio.controller.dto.response;
 
 import lombok.AccessLevel;
 import lombok.Builder;
-import org.sopt.api.stock.controller.dto.response.GetRelatedStockResponseDto;
 import org.sopt.domain.Stock;
 
 @Builder(access = AccessLevel.PRIVATE)
-public record GetPortfolioDto(
+public record GetPortfolioResponseDto(
         String code,
         String name,
         String currentPrice,
         float fluctuationRate
 ) {
-    public static GetPortfolioDto of(Stock stock) {
-        return GetPortfolioDto.builder()
+    public static GetPortfolioResponseDto of(Stock stock) {
+        return GetPortfolioResponseDto.builder()
                 .code(stock.getCode())
                 .name(stock.getName())
                 .currentPrice(stock.getCurrentPrice())
