+++ src/main/java/org/sopt/api/portfolio/service/PortfolioService.java
@@ -1,9 +1,8 @@
 package org.sopt.api.portfolio.service;
 
 import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.sopt.api.portfolio.controller.dto.GetPortfolioDto;
+import org.sopt.api.portfolio.controller.dto.response.GetPortfolioResponseDto;
 import org.sopt.domain.Portfolio;
 import org.sopt.domain.Stock;
 import org.sopt.domain.repository.PortfolioRepository;
@@ -18,10 +17,10 @@ public class PortfolioService {
     private final PortfolioRepository portfolioRepository;
     private final StockRepository stockRepository;
 
-    public List<GetPortfolioDto> getPortfolio(){
-        Portfolio portfolio = portfolioRepository.findByIdOrElseThrowException(Long.valueOf(1));
+    public List<GetPortfolioResponseDto> getPortfolio(){
+        Portfolio portfolio = portfolioRepository.findByIdOrElseThrowException(1L);
         List<Stock> stockList = stockRepository.findAllByPortfolio(portfolio);
 
-        return stockList.stream().map(stock -> GetPortfolioDto.of(stock)).toList();
+        return stockList.stream().map(stock -> GetPortfolioResponseDto.of(stock)).toList();
     }
 }
